pyproject.toml
[project]
name = "marketing-pipeline"
version = "0.1.0"
description = "Problem-first, project-agnostic marketing pipeline for Robert's open source tools."
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "atproto>=0.0.50",
    "Mastodon.py>=1.8.0",
    "pydantic>=2.6.0",
    "ruamel.yaml>=0.18.0",
    "httpx>=0.27.0",
    "tweepy>=4.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.5.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
marketing = "pipeline.cli:main"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["pipeline*"]
